refactor: move cogify up to top level
We'll reuse this function for other CDRs.
